pyproject.toml
[project]
name = "ndir-mocked"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt==1.6.1",
]
